requirements.txt
PyJWT
cryptography
requests
python-dotenv